build: fix compile warnings
To be able to spot future introduction of compile time issue, clean the
current state. To achieve this, we mainly change the type of variables
to match the data structures they are used for, add some initialization,
as well as reorder initializations.

+++ minisat/core/SolverTypes.h
@@ -272,7 +272,7 @@ class Clause
     /// remove the literal at the given position
     void remove_lit(uint32_t pos)
     {
-        assert(pos < size());
+        assert(pos < (size_t)size());
         data[pos].lit = last();
         pop();
     }
